// pronti-frontend-main/front/src/app/modulos/asiento-contable-cabecera/form/form-asiento-contable-cabecera.enum.ts
export enum FormAsientoContableCabeceraEnum {
  // nombre = 'nombre',
  fecha = 'fecha',
  anio = 'anio',
  mesPeriodo = 'mesPeriodo',
  tipoTransaccion = 'tipoTransaccion',
  tipoAsientoContable = 'tipoAsientoContable',
  codigoReferenciaAsientoContable = 'codigoReferencialAsientoContable',
  beneficiario = 'beneficiario',
  totalDebito = 'totalDebito',
  totalCredito = 'totalCredito',
  totalSaldoActualFecha = 'totalSaldoActualFecha',
  asientoCerrado = 'asientoCerrado',
  cheque = 'idCheque',
  serie = 'serie',
  descripcion = 'descripcion'

}